chore: move error type into error.rs
