+++ rust/dialog-remote-s3/src/lib.rs
@@ -66,11 +66,11 @@ pub use authorized::Authorized;
 pub use capability::{Access, Acl, Precondition};
 pub use capability::{archive, memory};
 pub use checksum::*;
-pub use error::AccessError;
+pub use error::{AccessError, S3StorageError};
 pub use permit::Permit;
 
 // Re-export site types at crate root
-pub use s3::{Bucket, RequestDescriptorExt, S3, S3StorageError};
+pub use s3::{Bucket, RequestDescriptorExt, S3};
 
 // Re-export S3Credentials at crate root for convenience
 pub use s3::S3Credentials;
